TracForgePlugin:
Stop trying to preserve the WSGI stream here; it doesn't work
reliably.  As far as we can tell, this requires a Trac patch as shown
in http://trac.edgewall.org/ticket/5697#comment:3

Although Noah claims this can be handled in a pre-request filter, he
hasn't been very forthcoming with details and says he won't have time
to work on it in the near future, so no breath held over here.