WikiCalendarMacro: Add configurable check logic.
Recently I've been facing a particularly tricky demand regarding
WikiTicketCalendarMacro that led to this by no means finished approach on
a general purpose check framework for calendar macros.

The first check I've been using this new code with is special as it's based
on a SQL query to a TracFormsPlugin db table, assuming there is a !TracForm
in every page defined by the wiki calendar, i.e. by means of a suitable
wiki page template.

I think of refactoring the currently 'embedded' checks like the basic one
for page existance and others for milestone and ticket create/due dates.
Expect more work in this area.